[smc] simplify
