+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auton/StateMachineTest.java
@@ -8,8 +8,6 @@ import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.subsystems.fsm.Edge;
 import org.firstinspires.ftc.teamcode.subsystems.fsm.State;
 import org.firstinspires.ftc.teamcode.subsystems.fsm.StateMachine;
-import org.firstinspires.ftc.teamcode.subsystems.fsm.StateMeta;
-import org.firstinspires.ftc.teamcode.subsystems.fsm.StateRole;
 
 /**
  * Autonomous OpMode to test the FSM subsystem.
@@ -41,7 +39,7 @@ public class StateMachineTest extends OpMode {
 /**
  * A state representing moving forward.
  */
-@StateMeta(color = "#E09F3E")
+@State.Meta(color = "#E09F3E")
 class ForwardState implements State {
     @Override
     public Edge<?>[] getEdges() {
@@ -61,7 +59,7 @@ class ForwardState implements State {
 /**
  * A state representing idelation.
  */
-@StateMeta(role = StateRole.INITIAL, color = "#335C67")
+@State.Meta(role = State.Role.INITIAL, color = "#335C67")
 class IdleState implements State {
     @Override
     public Edge<?>[] getEdges() {
