+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/fsm/State.java
@@ -1,9 +1,41 @@
 package org.firstinspires.ftc.teamcode.subsystems.fsm;
 
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
 /**
  * A state, belonging to an abstract state machine.
  */
 public interface State {
+    /**
+     * How to use the state; initial, transitional, or terminating.
+     */
+    enum Role {
+        /** This is the first state to jump to. */
+        INITIAL,
+
+        /** This state is nothing special in the eyes of the controller. */
+        TRANSITIONAL,
+
+        /** Once this state is run once, the state machine dies. */
+        TERMINATING
+    }
+
+    /**
+     * Describes some metadata about a state.
+     */
+    @Retention(RetentionPolicy.RUNTIME)
+    @Target(ElementType.TYPE)
+    @interface Meta {
+        /** The color to use when graphing with GraphViz DOT. */
+        String color() default "#000000";
+
+        /** How to use the state. */
+        Role role() default Role.TRANSITIONAL;
+    }
+
     /** The edges of the current state. */
     Edge<?>[] getEdges();
