--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/fsm/StateRole.java
@@ -1,15 +1,0 @@
-package org.firstinspires.ftc.teamcode.subsystems.fsm;
-
-/**
- * How to use the state; initial, transitional, or terminating.
- */
-public enum StateRole {
-    /** This is the first state to jump to. */
-    INITIAL,
-
-    /** This state is nothing special in the eyes of the controller. */
-    TRANSITIONAL,
-
-    /** Once this state is run once, the state machine dies. */
-    TERMINATING
-}
